day 11 part 2
1 hour overtime, but given the available time I had to work on this
today, I'm not too bothered by that.

The really frustrating thing is that I fundamentally understood the
proper solution for part2, but went down a long rabbit hole trying
to debug the overflow error where I assumed it was happening instead
of just looking at the backtrace to find out where it really happened.